requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
django-filter==25.1
Pillow==11.1.0
coreapi==2.3.3
psycopg2-binary==2.9.10 